fix: collapsed link 클릭 시 안 닫히는 버그 해결
