+++ src/Components/CollapseLinks/index.jsx
@@ -3,32 +3,12 @@ import { useRef, useEffect } from 'react';
 import { NavLink } from 'react-router-dom';
 
 const CollapseLinks = props => {
-    const { id = '', isOpened, items = [], closeTask } = props;
-    const ref = useRef();
-
-    const close = () => {
-        closeTask(state => !state);
-    };
-
-    const handleMouseDown = e => {
-        if (!isOpened) return;
-        if (ref.current.contains(e.target)) return;
-
-        close();
-    };
-
-    useEffect(() => {
-        document.addEventListener('mousedown', handleMouseDown);
-
-        return () => {
-            document.removeEventListener('mousedown', handleMouseDown);
-        };
-    }, [isOpened]);
+    const { id = '', isOpened, items = [] } = props;
 
     return (
-        <div id={id} ref={ref}>
+        <div id={id}>
             <div className={'collapse ' + (isOpened ? 'open' : '')}>
-                <div className='collapseInner' onClick={close}>
+                <div className='collapseInner'>
                     {items.map((item, index) => {
                         return (
                             <NavLink key={index} className='collapseItem' to={item.path}>
